+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 ## Changed
+- [#7](https://github.com/giuseppemorelli/digital-ocean-backup-systems/issues/7) snapshot exclude list moved to .env file 
 - [#8](https://github.com/giuseppemorelli/digital-ocean-backup-systems/issues/8) droplet exclude list moved to .env file 
 
 ## 0.2.0